Recreate content blocker package to remove sub library. Clean up logic and use @main executable
