+++ content-blocker-lib-ios/ContentBlockerGenerator/README.md
@@ -1,0 +1,3 @@
+# ContentBlockerGenerator
+
+TODO: A description of this package, what it does, how it's used.
